requirements/base.txt
# Core Framework
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20

# Database
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
alembic==1.14.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Data Validation
pydantic==2.10.6
pydantic-settings==2.7.1

# Task Queue
celery==5.4.0
redis==5.2.1

# Computer Vision & ML (Python 3.14 compatible)
opencv-python==4.11.0.86
ultralytics==8.3.80
torch==2.6.0
torchvision==0.21.0

# Image Processing
Pillow==11.1.0
numpy==2.2.3

# Reporting
reportlab==4.3.1
pandas==2.2.3

# File Handling
python-magic==0.4.27

# Utilities
requests==2.32.3
python-dateutil==2.9.0